fix(components/help-inline): remove dependency on indicators package
