+++ libs/components/help-inline/package.json
@@ -21,7 +21,6 @@
     "@angular/core": "^17.3.4",
     "@skyux/core": "0.0.0-PLACEHOLDER",
     "@skyux/i18n": "0.0.0-PLACEHOLDER",
-    "@skyux/indicators": "0.0.0-PLACEHOLDER",
     "@skyux/popovers": "0.0.0-PLACEHOLDER",
     "@skyux/theme": "0.0.0-PLACEHOLDER"
   },
