+++ libs/components/help-inline/src/lib/modules/help-inline/help-inline.component.scss
@@ -3,12 +3,16 @@
 @use 'libs/components/theme/src/lib/styles/_public-api/themes/modern/_compat/mixins'
   as modernMixins;
 
+:host {
+  display: inline-block;
+}
+
 .sky-help-inline {
   color: $sky-background-color-primary-dark;
   font-size: $sky-font-size-base;
   background-color: transparent;
   border: none;
-  display: inline-block;
+  display: inline-flex;
 
   &:hover {
     color: darken($sky-background-color-primary-dark, 10%);
@@ -22,19 +26,27 @@
 }
 
 @include mixins.sky-theme-modern {
-  sky-icon-stack {
+  .sky-icon-stack {
+    --size: 1.25em;
     display: inline-flex;
-  }
-  .sky-help-inline {
-    // The button mixin sets the border radius to 6px but we only want 3px on help inline buttons.
-    border-radius: 3px;
-    ::ng-deep .fa-stack-2x {
+    width: var(--size);
+    height: var(--size);
+    line-height: var(--size);
+
+    .fa-stack-2x {
       font-size: 16px;
     }
 
-    ::ng-deep .fa-stack-1x {
+    .fa-stack-1x {
       font-size: 10px;
     }
+  }
+  .sky-help-inline {
+    position: relative;
+    top: 0.15em;
+
+    // The button mixin sets the border radius to 6px but we only want 3px on help inline buttons.
+    border-radius: 3px;
 
     // The 0px and 5px padding is because we want 1px top/bottom and 6px left/right but the mixin
     // adds a pixel to account for standard button drop shadows that do not exist on this button.
