+++ libs/components/help-inline/src/lib/modules/help-inline/help-inline.component.spec.ts
@@ -228,27 +228,6 @@ describe('Help inline component', () => {
       );
     });
 
-    it('should use sky-icon in default theme', async () => {
-      fixture.detectChanges();
-
-      expect(fixture.nativeElement.querySelector('sky-icon')).toExist();
-      expect(fixture.nativeElement.querySelector('sky-icon-stack')).toBe(null);
-    });
-
-    it('should use sky-icon-stack in modern', async () => {
-      mockThemeSvc.settingsChange.next({
-        currentSettings: new SkyThemeSettings(
-          SkyTheme.presets.modern,
-          SkyThemeMode.presets.light,
-        ),
-        previousSettings: mockThemeSvc.settingsChange.value.currentSettings,
-      });
-      fixture.detectChanges();
-
-      expect(fixture.nativeElement.querySelector('sky-icon')).toBeNull();
-      expect(fixture.nativeElement.querySelector('sky-icon-stack')).toExist();
-    });
-
     it('should set aria label with labelText', async () => {
       component.labelText = 'test component';
